+++ spec/unit/test_relation.rb
@@ -194,6 +194,27 @@ module Bmg
       end
     end
 
+    describe 'restrict' do
+      let(:relation) {
+        Relation.new([
+          { a: 1, b: 2 },
+          { a: 2, b: 4 }
+        ])
+      }
+
+      subject {
+        relation.restrict(a: 1)
+      }
+
+      it_behaves_like "an operator method"
+
+      it 'returns the exected result' do
+        expect(subject.to_a).to eql([
+          { a: 1, b: 2 }
+        ])
+      end
+    end
+
     describe 'union' do
       let(:left) {
         Relation.new([